cleaned up code in admin set new components, AdminUser, OptionStates and OptionAvatars to increase readablity
